Add turbo encabulator
An instrument that allows transmission that not only supplies inverse
reactive current for use in unilateral phase detractors, but is also
capable of automatically synchronizing cardinal grammeters.

Basically, the only new principle involved is that instead of power
being generated by the relative motions of conductors and fluxes, it is
produced by the modial interaction of magnetoreluctance and capacitive
directance.

The original machine had a baseplate of prefabulated Amulite, surmounted
by a malleable logarithmic casing in such a way that the two spurving
bearings were in a direct line with the pentametric fan. The latter
consisted simply of six hydrocoptic marzel vanes so-fitted to the
ambifacient lunar wane shaft that side fumbling was effectively
prevented.

The main winding was of the normal lotus-o-delta type placed in
panendermic semi-boloid slots in the stator, every seventh conductor
being connected by a nonreversible trem'e pipe to the differential
girdlespring on the up-end of the grammeters.

The Turbo Encabulator has now reached a high level of develement, and is
being successfully used in the operation of nofer trunnions. Moreover,
whenever a barescent skor motion is required, it may be employed in
conjunction with a drawn reciprocating dingle arm to reduce sinusodial
depleneration.